Some cleanup in FRSystemProfile
- use long long instead of long for memory size and cpu speed
- replace some deprecated constants with newer names
- replace some magic numbers with named constants

+++ Sources/Main/FRSystemProfile.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2008, Torsten Curdt
+ * Copyright 2008-2014, Torsten Curdt
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -28,7 +28,7 @@
 + (int) cpucount;
 + (NSString*) machinemodel;
 + (NSString*) language;
-+ (long) cpuspeed;
-+ (long) ramsize;
++ (long long) cpuspeed;
++ (long long) ramsize;
 
 @end
